requirements.txt
openai
langchain
python-dotenv
faiss-cpu
chromadb
networkx
lxml
streamlit
tqdm
langchain
langchain-community
langchain-openai
